Issue#1 add home page Closes #1 (#6)
* new: setup base styles for the app

I am using Roboto font for the application. Stylesheet link
is added in the index.html page.

* update: Add styles for the container

* new(_mixins.scss): add mixin for media-query for the large devices

* new(api.service): add new service to interact with APIs

Add a function to fetch the news from the server.

* new(utility.service): add new service to provide utility functions

Add functions to shortened the URL and to get the formatted date.

* update(routes.jsx): add base route for the app

* update(index.jsx): wrap the App component with HOC to user routing feature

* new(Button.component): add new button component

* new(NavBar.component): add new NavBar component

Add the NavBar in the app component.

* new(NewsItem.component): add new NewsItem component

This component will render the news item provided the caption and
value.

* new(NewsDetails): add new NewsDetails component

This compoent will render the news details in a specific manner
provided in designs. This component will combine the below properties together:
1. Title of the news
2. URL of the news
3. Author of the news
4. Date of the news

* new(NewsCard.component): add a NewsCard component

This is a responsive card for displaying all the details of a news and
the actions to be taken.

Each news card will have two options:
1. Upvote
2. Hide

* nwe(NewsPage): add the NewsPage to render the news

* update(package.json): add new packages

* update(package.json): add pre-commit and pre-push hooks

* fix: test passing

Co-authored-by: Kanishka Mohan Madhuni <madhuni@Kanishkas-MacBook-Air.local>